feat： 完善角色接口

+++ src/entity/role_access_entity.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct RoleAccessEntity {
-    pub id: Option<u16>,
+    pub id: Option<i32>,
     pub role_id: i32,
     pub access_id: i32,
 }
